refactor: group market components
