+++ src/market/MarketHelper.tsx
@@ -1,25 +1,25 @@
 import { useState, useEffect, useMemo, useRef, memo, useContext } from "react";
 import { invoke } from "@tauri-apps/api/core";
-import { ImgCacheDirContext } from "./ImgCacheDir";
+import { ImgCacheDirContext } from "../ImgCacheDir";
 import { listen } from "@tauri-apps/api/event";
-import { HelpTip } from "./shared/HelpTip";
+import { HelpTip } from "../shared/HelpTip";
 import WfmTrading from "./WfmTrading";
 import ItemMarketPopup from "./ItemMarketPopup";
-import { warframeStatImageUrl } from "./constants/urls";
-import { MARKET_FILTERS_DEFAULT } from "./constants/filters";
-import { TAURI_COMMANDS } from "./constants/tauri";
-import type { CatalogItem, CraftingJob, InventoryItem, RecipeComponent, RecipeMap } from "./types/items";
-import type { MarketFilters } from "./types/filters";
-import type { ModCopy } from "./types/inventory";
-import type { BlobRivenEntry, BlobRivenStat, WfmCachedPrices, WfmItem, WfmItemInfo, WfmPrice, WfmPriceUpdate, WfmRivenAttribute } from "./types/market";
-import type { WfmCreateOrderArgs, WfmCreateRivenAuctionArgs, WfmSession } from "./types/tauri";
-import polMadurai  from "./assets/polarity/madurai.svg";
-import polVazarin  from "./assets/polarity/vazarin.svg";
-import polNaramon  from "./assets/polarity/naramon.svg";
-import polZenurik  from "./assets/polarity/zenurik.svg";
-import polUnairu   from "./assets/polarity/unairu.svg";
-import polPenjaga  from "./assets/polarity/penjaga.svg";
-import polUmbra    from "./assets/polarity/umbra.svg";
+import { warframeStatImageUrl } from "../constants/urls";
+import { MARKET_FILTERS_DEFAULT } from "../constants/filters";
+import { TAURI_COMMANDS } from "../constants/tauri";
+import type { CatalogItem, CraftingJob, InventoryItem, RecipeComponent, RecipeMap } from "../types/items";
+import type { MarketFilters } from "../types/filters";
+import type { ModCopy } from "../types/inventory";
+import type { BlobRivenEntry, BlobRivenStat, WfmCachedPrices, WfmItem, WfmItemInfo, WfmPrice, WfmPriceUpdate, WfmRivenAttribute } from "../types/market";
+import type { WfmCreateOrderArgs, WfmCreateRivenAuctionArgs, WfmSession } from "../types/tauri";
+import polMadurai  from "../assets/polarity/madurai.svg";
+import polVazarin  from "../assets/polarity/vazarin.svg";
+import polNaramon  from "../assets/polarity/naramon.svg";
+import polZenurik  from "../assets/polarity/zenurik.svg";
+import polUnairu   from "../assets/polarity/unairu.svg";
+import polPenjaga  from "../assets/polarity/penjaga.svg";
+import polUmbra    from "../assets/polarity/umbra.svg";
 
 interface Props {
   inventory: Record<string, InventoryItem>;
